Adding support for custom geoJson layers
Some leaflet plugins such as Proj4Leaflet, and Leaflet.GeoJSON.Encoded extend the standard L.GeoJSON class to support features such as custom coordinate reference systems, and encoding styles.  This modification is designed to allow this plugin to work in coordination with other plugins which may have a customized version of the L.geoJson by accepting the layer type as an option with the L.geoJson layer type the default option.